% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archr2MAE.R
\name{archr2MAE}
\alias{archr2MAE}
\title{Create a MultiAssayExperiment from an ArchR project}
\usage{
archr2MAE(archrDir, numThreads = 1)
}
\arguments{
\item{archrDir}{String specifying the ArchR project directory. The directory should contain a Save-ArchR-Project.rds.}

\item{numThreads}{Integer specifying threads to be used for parallel computing during ArchR project setup}
}
\value{
A \linkS4class{MultiAssayExperiment}
}
\description{
Import the matrices from a ArchR project to into a MultiAssayExperiment where each ArchR matrix is a SingleCellExperiment. Metadata is not added to the MultiAssayExperiment or Experiments.
}
\author{
Natalie Fox
}
