% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSCE.R
\name{convertSCE}
\alias{convertSCE}
\title{convert SingleCellExperiment to custom class}
\usage{
convertSCE(sce, newClass)
}
\arguments{
\item{sce}{a \code{SingleCellExperiment}}

\item{newClass}{character string specifying the target class}
}
\value{
Object of class \code{newClass}. If conversion fails, the original \code{sce}.
}
\description{
Converts an SCE object to a custom class, if a method is defined.
}
\details{
This function attempts to convert a SingleCellExperiment object to another class.
A class-specific method for \code{newClass} is looked for in this package's namespace.
The method name \strong{must} follow the format: "convertTo\code{<newClass>}".
If no such method is found, the default is to use \code{as}.
If that fails, the original \code{SingleCellExperiment} will be returned.
}
\seealso{
\code{loadExp}
}
