% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRawMatrixSeed.R
\name{createRawMatrixSeed}
\alias{createRawMatrixSeed}
\alias{.createRawMatrixSeed}
\alias{.extractMatrixDimnames}
\title{Matrix loading utilities}
\usage{
.createRawMatrixSeed(info, path, names = TRUE)

.extractMatrixDimnames(path)
}
\arguments{
\item{info}{A named list of metadata for this matrix.}

\item{path}{String containing the path to the file containing said matrix.}

\item{names}{Logical scalar indicating whether the seed should be annotated with dimnames (if available).}
}
\value{
\code{.createRawMatrixSeed} returns a seed that can be used in the \code{DelayedArray} constructor.

\code{.extractMatrixDimnames} returns a list of two character vectors or \code{NULL}, containing the dimnames.
}
\description{
Utilities for loading a matrix saved by \code{stageObject}.
}
\details{
For \code{.extractDimnames}, \code{path} is expected to be a HDF5 file
with row and column names in \code{samp_data/features} and \code{samp_data/samples}, respectively.
}
\section{Note}{

This function has been kindly contributed by Aaron Lun.
}

\author{
Aaron Lun
}
\keyword{internal}
