% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classDefinitions.R
\docType{class}
\name{customClasses}
\alias{customClasses}
\title{custom classes}
\description{
Additional class definitions.
}
\section{Classes}{

\itemize{
\item{\code{SingleCellAccessibilityExperiment}, contains \code{SingleCellExperiment}}
}
}

\keyword{internal}
