% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineDataset.R
\name{determineDataset}
\alias{determineDataset}
\title{determine dataset being queried}
\usage{
determineDataset()
}
\value{
A character string reflecting the name of the dataset.
}
\description{
Obtain name of dataset from the (grand)parent call.
}
\details{
This function reads the name of the function two steps up the call stack and
returns is as a string. It is called by \code{loadExp} and \code{getMetadata}
in order to determine the multiome dataset being queried.
}
\keyword{internal}
