% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummies.R
\name{dummies}
\alias{dummies}
\alias{dummySCE}
\alias{dummyMAE}
\title{create dummy data sets}
\usage{
dummySCE(...)

dummyMAE(experiments = c("EXP1", "EXP2"))
}
\arguments{
\item{experiments}{character string of experiment names}
}
\value{
\code{dummySCE} returns a \code{SingleCellExperiment}.
\code{dummyMAE} returns a \code{MultiAssayExperiment}.
}
\description{
Create summy SCE and MAE objects
}
\examples{
scMultiome:::dummySCE()
scMultiome:::dummyMAE("dummyExperiment")

}
\keyword{internal}
