% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSCE.R
\name{experimentClass}
\alias{experimentClass}
\alias{convergeSCE}
\alias{convertSCE}
\title{convert custom experiments classes to and from SingleCellExperiment}
\usage{
convergeSCE(exp, customClass)

convertSCE(sce, customClass)
}
\arguments{
\item{exp}{an experiment other than a \code{SingleCellExperiment}}

\item{customClass}{character string specifying the custom class}

\item{sce}{a \code{SingleCellExperiment}}
}
\value{
\code{convergeSCE} returns an object of class \code{SingleCellExperiment}
or an error if conversion fails.
\code{convertSCE} returns an object of class \code{customClass}
or the original \code{sce} if conversion fails.
}
\description{
Converts a custom class experiment object to SCE and \emph{vice versa},
if methods are defined.
}
\details{
Multiome data sets are saved as MultiAssayExperiments composed of SingleCellExperiments.
If a data set contains other classes, they must be converted to SCE before saving
and reconverted to those classes after loading. Some packages provide handlers
for such conversions, either utilizing the \code{methods::as} generic,
or defining their own methods.
These functions provide a catch-all mechanism to do the conversions automatically
during saving and loading experiments.
Authors adding new data sets may choose to do it manually.

\code{convergeSCE} is called by \code{saveExp}.
It attempts to convert an experiment of another class to SingleCellExperiment.
A class-specific method is searched for in this package's namespace.
This method \strong{must} follow the naming format: "convertFrom\code{<customClass>}".
If no such method is found, the default is to use \code{as}.
If that fails, an error is signaled.

\code{convertSCE} is called by \code{loadExp}.
It attempts to convert a SingleCellExperiment to another class.
A class-specific method is search for in this package's namespace.
This method \strong{must} follow the naming format: "convertTo\code{<customClass>}".
If no such method is found, the default is to use \code{as}.
If that fails, the original \code{SingleCellExperiment} will be returned.
}
\seealso{
\code{saveExp}, \code{loadExp}
}
\keyword{internal}
