% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileUtils.R
\name{fileUtils}
\alias{fileUtils}
\alias{testFile}
\alias{uploadFile}
\title{file utilities}
\usage{
testFile(file)

uploadFile(
  file,
  sasToken,
  endpoint = "https://bioconductorhubs.blob.core.windows.net"
)
}
\arguments{
\item{file}{path to hdf5 file containing a data set}

\item{sasToken}{access token to \code{endpoint}}

\item{endpoint}{Bioconductor's data bucket endpoint url}
}
\value{
\code{testFile} returns the \code{MultiAssayExperiment} stored in \code{file} in its entirety.
\code{uploadFile} returns TRUE invisibly.
}
\description{
Helper functions for new data sets.
}
\details{
\code{testFile} can be used to test whether a data set loads correctly from a local file.
It calls \code{loadMAE} and extracts all experiment verbosely.

\code{uploadFile} will load a single file to the Bioconductor staging directory.

Both functions are internal as they are meant to be run in developer mode.
}
\keyword{internal}
