% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step6b_FinalizeExprData.R
\name{FinalizeExpressionData}
\alias{FinalizeExpressionData}
\title{Finalize Expression Data by Filtering for Protein-Coding Genes}
\usage{
FinalizeExpressionData(expr_gr, protein_coding_gr)
}
\arguments{
\item{expr_gr}{A \code{GRanges} object of expression data, with gene IDs in metadata column \code{Expr_id}.}

\item{protein_coding_gr}{A \code{GRanges} object of protein-coding genes, with \code{gene_id} and \code{gene_name} in metadata.}
}
\value{
A filtered \code{GRanges} object containing only protein-coding genes,
with an added metadata column \code{Gene_Symbol}.
}
\description{
This function performs Step 6 of the exam:
takes a GRanges object of expression data and filters it to retain
only those entries that correspond to protein-coding genes (based on GTF annotation).
It also adds gene symbol names as a metadata column for easier interpretation.
}
\examples{
\dontrun{
expr_filtered <- FinalizeExpressionData(expr_gr, protein_coding_gr)
}

}
