% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{getMetadata}
\alias{getMetadata}
\title{retrieve data set metadata}
\usage{
getMetadata()
}
\value{
An \code{ExperimentHub} object.
}
\description{
Retrieve and return metadata for the currently queried dataset in ExperimentHub format.
}
\details{
This is an internal function that is used by accessor functions when metadata is requested.
Dataset name is derived from the name of the accessor function which calls this one.
}
