% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRtools.R
\name{RGtools}
\alias{RGtools}
\alias{storeGR}
\alias{restoreGR}
\title{manipulate GeneRanges}
\usage{
storeGR(x)

restoreGR(df)
}
\arguments{
\item{x}{object of class \code{GRanges}}

\item{df}{a \code{data.frame}}
}
\value{
\code{storeGR} returns a data frame, \code{restoreGR} returns a \code{GRanges} object.
}
\description{
Prepare and recover \code{GRanges} objects for and after storing.
}
\details{
\code{GRanges} objects, which can be encountered in \code{rowRanges} slots of
\code{SingleCellExperiment}s, are stored as data frames (of type compound).

\code{storeGR} converts \code{GRanges} to a data frames converts factors to characters.
\code{restoreGR} resets data types in the basic columns and re-instantiates \code{GRanges}.
}
\keyword{internal}
