% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDatasets.R
\name{listDatasets}
\alias{listDatasets}
\title{list all available datasets}
\usage{
listDatasets()
}
\value{
A \code{DataFrame} listing all available datasets,
with one datasest per row and the following columns:
\itemize{
\item{Species: species name}
\item{Type: sample type, e.g. cell culture, tissue}
\item{Multiome: paired or unpaired}
\item{Call: function call used to access the dataset directly}
}
}
\description{
Summary information for all datasets available in the package.
}
