% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDataSetList.R
\name{makeDataSetList}
\alias{makeDataSetList}
\title{create data set list}
\usage{
makeDataSetList(metadata)
}
\arguments{
\item{metadata}{a \code{data.frame} containing data set metadata}
}
\value{
Invisible TRUE.
}
\description{
Automatically creates the data set list of the package.
}
\details{
This is an internal helper function for developers and will not be called directly.
It creates the file inst/scripts/datasetList.Rmd, which is incorporated
into the package help page to automatically list the current data sets.
}
\keyword{internal}
