% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMakeData.R
\name{makeMakeData}
\alias{makeMakeData}
\title{create make-data file}
\usage{
makeMakeData(dataset)
}
\arguments{
\item{dataset}{name of data set as character string}
}
\value{
Invisible TRUE.
Creates \code{inst/scripts/make-data-<dataset>.Rmd} file and navigates to it, if possible.
}
\description{
Creates a template for data set history file.
}
