% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMakeMetadata.R
\name{makeMakeMetadata}
\alias{makeMakeMetadata}
\title{create make-metadata data file}
\usage{
makeMakeMetadata(dataset)
}
\arguments{
\item{dataset}{name of data set as character string}
}
\value{
Invisible TRUE.
Creates \code{inst/scripts/make-metadata-<dataset>.R} file and navigates to it, if possible.
}
\description{
Creates a template for data set metadata file.
}
