% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeR.R
\name{makeR}
\alias{makeR}
\title{create accessor function file}
\usage{
makeR(dataset)
}
\arguments{
\item{dataset}{name of data set as character string}
}
\value{
Invisible TRUE.
Creates \code{R/<dataset>.Rmd} file and navigates to it, if possible.
}
\description{
Creates a template for accessor function and data set help.
}
