% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{motifs}
\alias{motifs}
\title{Transcription Factor Motif Info}
\usage{
motifs(metadata = FALSE, genome = c("hg38", "hg19", "mm10"))
}
\arguments{
\item{metadata}{logical flag specifying whether to retrieve the data (FALSE) or metadata only (TRUE)}

\item{genome}{character string specifying the genome build}
}
\value{
A \code{GRangeList} object containing binding site information of transcription factor
ChIP-seq combined from Cistrome database and ENCODE.
}
\description{
Transcription factor binding motif information.
}
\details{
This is a special data set that stores transcription factor binding motifs
(for several genome builds), which can be used with multiome data to compute epiregulons.
}
\section{Data storage and access}{

All motif info objects (\code{GRangeList}) are stored in one hdf5 file.
The function will return the one specified in the call.
}

\section{Data preparation}{


Data were retrieved from (internal) \code{genomitory} data base.
}

\examples{
motifs("mm10")

}
