% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padZeros.R
\name{padZeros}
\alias{padZeros}
\title{pad numbers in strings with zeros}
\usage{
padZeros(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
A character where all elements have the same number of characters.
}
\description{
Insert "0" characters to unify number of characters in character vector.
}
\details{
Given a character vector composed of letters and digits,
add zeros at the beginning of the numeric part so that all strings are of the same length.
}
\examples{
padZeros(paste0("C", 1:5))
padZeros(paste0("C", 1:10))

}
