% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve}
\alias{retrieve}
\title{retrieve data set or its metadata}
\usage{
retrieve(dataset, metadata, experiments, verbose = FALSE)
}
\arguments{
\item{dataset}{character string specifying the data set name}

\item{metadata}{logical flag specifying whether to return the resource or only its metadata}

\item{experiments}{character string specifying which experiments to extract}

\item{verbose}{logical flag specifying loading verbosity}
}
\value{
If \code{metadata = FALSE}, a \code{MultiAssayExperiment},
otherwise an \code{ExperimentHub} object.
}
\description{
Retrieve and return the data or metadata for the currently queried dataset.
}
\details{
This is a generic accessor function that is used by user-level accessor functions
to access datasets or their metadata.
}
\keyword{internal}
