% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{templates}
\alias{templates}
\alias{makeMakeData}
\alias{makeMakeMetadata}
\alias{makeR}
\title{documentation templates}
\usage{
makeMakeData(dataset)

makeMakeMetadata(dataset)

makeR(dataset)
}
\arguments{
\item{dataset}{name of data set as character string}
}
\value{
All functions return TRUE invisibly.
}
\description{
Create templates for data set documentation.
}
\details{
Functions to facilitate documenting a new data set.
Each function creates a file and attempts to open it in RStudio for editing.

\code{makeMakeData} creates an Rmarkdown report called \code{inst/scripts/make-data-<dataset>.Rmd}.\cr
\code{makeMakeMetadata} creates an R script called \code{inst/scripts/make-metadata-<dataset>.R}.\cr
\code{makeR} creates an R file called \code{R/<dataset>.Rmd}.\cr
}
\author{
Aleksander Chlebowski
}
