% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfBinding_chipatlas_encode.R
\name{tfBinding_chipatlas_encode}
\alias{tfBinding_chipatlas_encode}
\title{TF Binding Info}
\format{
\code{GRangeList} object containing binding site information of transcription factor
ChIP-seq merged from ChIPAtlas database and ENCODE.
Contains the following experiments:
\itemize{
\item{\strong{hg38}: GRangesList object of length 1558}
\item{\strong{hg19}: GRangesList object of length 1558}
\item{\strong{mm10}: GRangesList object of length 768}
}
}
\usage{
tfBinding_chipatlas_encode(
  genome = c("hg38", "hg19", "mm10"),
  metadata = FALSE
)
}
\arguments{
\item{genome}{String indicating the genomic build}

\item{metadata}{logical flag specifying whether to return data or metadata only}
}
\value{
Granges List of TF binding sites
}
\description{
Combined transcription factor ChIP-seq data from ChIP-Atlas and ENCODE.
This is a special data set that stores transcription factor binding sites for human and
mouse genomic builds, which can be used with multiome data to compute epiregulons.
}
\section{Data preparation}{

}

\section{Data storage and access}{

The \code{MultiAssayExperiments} is split into separate \code{SingleCellExperiment}
objects and they in turn are split into components, all of which are stored in a
single hdf5 file. Data and can be accessed with a special function that extracts
elements of the requested experiment(s), reassembles them, and builds an MAE.
}

\examples{
tfBinding_chipatlas_encode("hg38")

}
\references{
ChIP-Atlas 2021 update: a data-mining suite for exploring epigenomic landscapes by
fully integrating ChIP-seq, ATAC-seq and Bisulfite-seq data.
Zou Z, Ohta T, Miura F, Oki S.
\emph{Nucleic Acids Research. Oxford University Press (OUP);} 2022.
\href{http://dx.doi.org/10.1093/nar/gkac199}{doi:10.1093/nar/gkac199}

ChIP‐Atlas: a data‐mining suite powered by full integration of public ChIP‐seq data.
Oki S, Ohta T, Shioi G, Hatanaka H, Ogasawara O, Okuda Y, Kawaji H, Nakaki R, Sese J, Meno C.
\emph{EMBO}; Vol. 19, EMBO reports. 2018.
\href{http://dx.doi.org/10.15252/embr.201846255}{doi: 10.15252/embr.201846255}

ENCODE {https://www.encodeproject.org/}
}
