% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummies.R
\name{dummies}
\alias{dummies}
\alias{dummySCE}
\alias{dummyMAE}
\title{create dummy data sets}
\usage{
dummySCE(
  features = c("rowData", "rowRanges", "reducedDims", "altExps", "none")
)

dummyMAE(experiments = list(EXP1 = NULL, EXP2 = NULL))
}
\arguments{
\item{features}{character string specifying which (optional) features to create in the SCE}

\item{experiments}{named list of character vectors
specifying experiments to create and their features}
}
\value{
\code{dummySCE} returns a \code{SingleCellExperiment}.
\code{dummyMAE} returns a \code{MultiAssayExperiment}.
}
\description{
Create dummy SCE and MAE objects.
}
\examples{
scMultiome:::dummySCE()
scMultiome:::dummySCE("rowData")
scMultiome:::dummySCE("rowRanges")
scMultiome:::dummySCE("reducedDims")
scMultiome:::dummySCE("altExps")

scMultiome:::dummyMAE(list("dummyExperiment" = NULL))

}
\keyword{internal}
