% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMultiome-package.R
\docType{package}
\name{scMultiome-package}
\alias{scMultiome}
\alias{scMultiome-package}
\title{scMultiome: Collection of Public Single-Cell Multiome (scATAC + scRNAseq) Datasets}
\description{
Single cell multiome data, containing chromatin accessibility (scATAC-seq) and gene expression (scRNA-seq) information analyzed with the ArchR package and presented as MultiAssayExperiment objects.
}
\details{
Single cell multiome data sets, paired and unpaired, were analyzed
with the \code{ArchR} package in order to obtain epiregulons.
\code{ArchR} projects were converted to \code{MultiAssayExperiment}
objects and serialized as .RDS files.
}
\section{Datasets}{

\itemize{
\item{\strong{prostateENZ}:
LNCaP cells cultured with or without enzalutamide.
}

}
}

\author{
\strong{Maintainer}: Xiaosai Yao \email{xyao19@gene.com}

Authors:
\itemize{
  \item Aleksander Chlebowski \email{aleksander.chlebowski@contractors.roche.com}
}

}
\keyword{internal}
